dev: Add `pkg/commands` doc comments
